adsAsynPortDriver.cpp Fix compiler warning when printing struct timeval
Fix a compiler warning:
struct timeva.tv_usec may be defined as int, unsigned int, or long.
In practice, on a 32 bit system,
these do all the same - since the value is under 2GiB anyway.

However, the compiler complains that the printf uses "%ld",
when an int is passed.
That is a legal complain.
Solution: cast it to long.

+++ adsApp/src/adsAsynPortDriver.cpp
@@ -71,7 +71,7 @@ static void getEpicsState(initHookState state)
     case initHookAfterInitDatabase:
         gettimeofday(&now, NULL);
         timersub(&now, &start, &diff);
-        printf("Database initialization took %ld.%05ld seconds.\n", diff.tv_sec, diff.tv_usec);
+        printf("Database initialization took %ld.%05ld seconds.\n", diff.tv_sec, (long)diff.tv_usec);
         break;
     case initHookAfterScanInit:
       allowCallbackEpicsState=1;
